Update skipped_packages
Most packages don't really require a bdist.
We used to use eggs but now we use wheels, which
the packages have on PyPI. We also used to need to recompile
Windows wheels to use static CRT but right now we
bring the CRT with us so we don't need that either.